+++ udpcom/NetCommunicatorMaster.cxx
@@ -456,6 +456,7 @@ void NetCommunicatorMaster::checkAndUpdatePeerStates(const TimeSpec &ts)
           correctFollowId((*pp)->send_id, (*pp)->follow_id);
           clientInfoPeerLeft((*pp)->send_id, ts);
           pp = peers.erase(pp);
+          npeers--;
           continue;
         }
         case Accept: {
@@ -665,7 +666,7 @@ void NetCommunicatorMaster::doCycle(const TimeSpec &ts, Activity &activity)
   if (npeers < peer_cycles.size()) {
     for (peer_cycles_type::iterator pp = peer_cycles.begin();
          pp != peer_cycles.end();) {
-      if (message_cycle.cycleIsCurrentOrPast(pp->second)) {
+      if (!pp->second.cycleIsUpToDate(message_cycle)) {
         peer_cycles_type::iterator toerase = pp;
         /* DUECA network.
 
@@ -798,6 +799,14 @@ void NetCommunicatorMaster::unpackPeerData(MessageBuffer::ptr_type &buffer)
         clientUnpackPayload(buffer, i_.peer_id, current_tick, i_.peertick,
                             i_.usecs_offset);
         peer_cycles[i_.peer_id] = message_cycle;
+
+        /* DUECA network.
+
+           Information message, a peer will be added at the specified
+           cycle.
+        */
+        W_NET("Adding peer " << i_.peer_id << " to tracking, at cycle "
+                             << i_.cycle);
       }
       else {
         /* DUECA network.